Rename registry-index to registry-query
