+++ cmd/registry-query/index/index.go
