+++ cmd/registry-query/index/operations.go
